DefaultLayout: remove unused code & move it
