+++ src/pages/_layouts/index.ts
@@ -1,4 +1,4 @@
-export { default as DefaultLayout } from '../../beethovenx/pages/_layouts/DefaultLayout.vue';
+export { default as DefaultLayout } from './DefaultLayout.vue';
 export { default as BlankLayout } from './BlankLayout.vue';
 export { default as PoolTransferLayout } from './PoolTransferLayout.vue';
 export { default as FocusedLayout } from './FocusedLayout.vue';
